Move pi_hole base entity to separate module
