+++ homeassistant/components/pi_hole/__init__.py
@@ -22,12 +22,7 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.exceptions import ConfigEntryAuthFailed
 from homeassistant.helpers import entity_registry as er
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
-from homeassistant.helpers.device_registry import DeviceInfo
-from homeassistant.helpers.update_coordinator import (
-    CoordinatorEntity,
-    DataUpdateCoordinator,
-    UpdateFailed,
-)
+from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
 from .const import CONF_STATISTICS_ONLY, DOMAIN, MIN_TIME_BETWEEN_UPDATES
 
@@ -140,35 +135,3 @@ async def async_setup_entry(hass: HomeAssistant, entry: PiHoleConfigEntry) -> bo
 async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Unload Pi-hole entry."""
     return await hass.config_entries.async_unload_platforms(entry, PLATFORMS)
-
-
-class PiHoleEntity(CoordinatorEntity[DataUpdateCoordinator[None]]):
-    """Representation of a Pi-hole entity."""
-
-    def __init__(
-        self,
-        api: Hole,
-        coordinator: DataUpdateCoordinator[None],
-        name: str,
-        server_unique_id: str,
-    ) -> None:
-        """Initialize a Pi-hole entity."""
-        super().__init__(coordinator)
-        self.api = api
-        self._name = name
-        self._server_unique_id = server_unique_id
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Return the device information of the entity."""
-        if self.api.tls:
-            config_url = f"https://{self.api.host}/{self.api.location}"
-        else:
-            config_url = f"http://{self.api.host}/{self.api.location}"
-
-        return DeviceInfo(
-            identifiers={(DOMAIN, self._server_unique_id)},
-            name=self._name,
-            manufacturer="Pi-hole",
-            configuration_url=config_url,
-        )
