+++ homeassistant/components/pi_hole/switch.py
@@ -14,8 +14,9 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers import config_validation as cv, entity_platform
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import PiHoleConfigEntry, PiHoleEntity
+from . import PiHoleConfigEntry
 from .const import SERVICE_DISABLE, SERVICE_DISABLE_ATTR_DURATION
+from .entity import PiHoleEntity
 
 _LOGGER = logging.getLogger(__name__)
